apps/chatbot-evaluate/pyproject.toml
[tool.poetry]
description = "Chatbot-evaluate repository for PagoPA"
name = "chatbot-evaluate"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
llama-index = "^0.14.8"
llama-index-embeddings-google-genai = "^0.3.1"
llama-index-llms-google-genai = "^0.7.3"
python = ">=3.12,<3.14"
pyyaml = "^6.0.2"
ragas = "^0.2.0"
toml = "^0.10.2"

[tool.poetry.group.test.dependencies]
httpx = "^0.28.1"
pytest = "^8.3.5"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
